+++ lib/lightning/workflows/edge.ex
@@ -91,6 +91,12 @@ defmodule Lightning.Workflows.Edge do
 
   def validate(changeset) do
     changeset
+    |> validate_uuid([
+      :id,
+      :source_job_id,
+      :source_trigger_id,
+      :target_job_id
+    ])
     |> assoc_constraint(:workflow)
     |> assoc_constraint(:source_trigger)
     |> assoc_constraint(:source_job)